+++ src/utils/countdownToDate/countdownToDate.types.ts
@@ -1,0 +1,8 @@
+interface ICountdownToDateReturn {
+  days: number;
+  hours: number;
+  minutes: number;
+  seconds: number;
+}
+
+export type { ICountdownToDateReturn };