+++ src/utils/countdownToDate/index.ts
@@ -1,0 +1,3 @@
+import countdownToDate from './countdownToDate';
+
+export default countdownToDate;