+++ src/utils/isAndroid/index.ts
@@ -1,0 +1,3 @@
+import isAndroid from './isAndroid';
+
+export default isAndroid;