+++ src/utils/isIE/isIE.ts
@@ -1,0 +1,10 @@
+/**
+ * Checks if the user is using Internet Explorer.
+ * @param {Window} window Window object.
+ * @returns {boolean} If the browser is Internet Explorer.
+ */
+function isIE(window: Window): boolean {
+  return /msie|trident/i.test(window?.navigator?.userAgent.toLowerCase());
+}
+
+export default isIE;