+++ src/vtex/getCollectionShelf/index.ts
@@ -1,0 +1,3 @@
+import getCollectionShelf from './getCollectionShelf';
+
+export default getCollectionShelf;