+++ src/vtex/getProductVariations/getProductVariations.test.ts
@@ -1,0 +1,9 @@
+import getProductVariations from './getProductVariations';
+
+describe('getProductVariations', () => {
+  it('should be able to get product variants', async () => {
+    const variants = await getProductVariations('econverse', 25);
+
+    expect(variants).toHaveProperty('skus');
+  });
+});