+++ src/vtex/index.ts
@@ -1,4 +1,15 @@
 import MasterData from './MasterData';
 import getProductById from './getProductById';
+import getProductSuggestions from './getProductSuggestions';
+import getProductVariations from './getProductVariations';
+import getCollectionShelf from './getCollectionShelf';
+import getProfile from './getProfile';
 
-export { MasterData, getProductById };
+export {
+  MasterData,
+  getProductById,
+  getProductSuggestions,
+  getProductVariations,
+  getCollectionShelf,
+  getProfile,
+};